+++ CallerDirectory/Migrations/CallingContextModelSnapshot.cs
@@ -21,15 +21,15 @@ namespace CallerDirectory.Migrations
 
             modelBuilder.Entity("CallerDirectory.Models.CallRecord", b =>
                 {
-                    b.Property<int>("Id")
+                    b.Property<long>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("int");
+                        .HasColumnType("bigint");
 
                     b.Property<long?>("Caller")
                         .HasColumnType("bigint");
 
-                    b.Property<float>("Cost")
-                        .HasColumnType("float");
+                    b.Property<int>("Cost")
+                        .HasColumnType("int");
 
                     b.Property<string>("Currency")
                         .IsRequired()
